Replace instrument check in QG gen with instrument-defaulting.
Using a Pipeline that contains an instrument entry now automatically
includes that instrument as a constraint in the QG gen query, and
daf_butler will check that it's consistent with the instrument provided
in the user expression (which after this change is no longer required).

This is on the branch for this ticket only because the daf_butler
changes for this ticket reshuffled some modules, which broke the code
here that was previously using those (internal to daf_butler)
interfaces.  Happily, this is better behavior, better encapsulation,
and less code (well, not overall, but here).

On the testing side, we can remove a bunch of tests here (the
functionality we delegate to in daf_butler is already tested there),
but we also need to make the testing utility code a bit less cavalier,
and in particular make the pipeline's instrument agree with the
registry's instrument.